requirements.txt
tensorflow-cpu
equinox
optax
flax
jax